+++ src/Exceptions/FileNotReadableException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace SandFox\Bencode\Exceptions;
 
-interface BencodeException extends \Throwable
+class FileNotReadableException extends RuntimeException
 {
 }