+++ tests/FileTest.php
@@ -7,6 +7,8 @@ namespace SandFox\Bencode\Tests;
 use org\bovigo\vfs\vfsStream;
 use PHPUnit\Framework\TestCase;
 use SandFox\Bencode\Bencode;
+use SandFox\Bencode\Exceptions\FileNotReadableException;
+use SandFox\Bencode\Exceptions\FileNotWritableException;
 use SandFox\Bencode\Exceptions\InvalidArgumentException;
 
 class FileTest extends TestCase
@@ -61,13 +63,22 @@ class FileTest extends TestCase
         self::assertEquals($loaded, $value);
     }
 
-    public function testInvalidFile(): void
+    public function testInvalidFileRead(): void
     {
+        $this->expectException(FileNotReadableException::class);
         vfsStream::setup();
         $stream = vfsStream::newFile('test', 0000);
 
-        @self::assertEquals(false, Bencode::dump([], $stream->url()));
-        @self::assertEquals(false, Bencode::load($stream->url()));
+        self::assertEquals(false, Bencode::load($stream->url()));
+    }
+
+    public function testInvalidFileWrite(): void
+    {
+        $this->expectException(FileNotWritableException::class);
+        vfsStream::setup();
+        $stream = vfsStream::newFile('test', 0000);
+
+        self::assertEquals(false, Bencode::dump([], $stream->url()));
     }
 
     public function testEncodeToInvalidResource(): void