+++ CHANGELOG.md
@@ -37,6 +37,13 @@
 
 ## 2.x
 
+### 2.7.3
+
+*Oct 24, 2021*
+
+* dump() now throws exception if the file is not writable
+* load() now throws exception if the file is not readable
+
 ### 2.7.2
 
 *Oct 23, 2021*
@@ -121,6 +128,13 @@
 * Encode now throws an error if it encounters a value that cannot be serialized
 
 ## 1.x
+
+### 1.7.3
+
+*Oct 24, 2021*
+
+* dump() now throws exception if the file is not writable
+* load() now throws exception if the file is not readable
 
 ### 1.7.2